Moving model related code into another file
